RavenDB-15776 - PR comments
